Fix unrelated file being interpreted as a test
